Define a car class.
The C# code that is provided defines a car class that has attributes like Model, Manufacturer, Year, and RentalPrice. It also includes a DisplayDetails() function that shows the details of the car. The Car class adds attributes unique to vehicles, such as Seats, EngineType, Transmission, and Convertible, after inheriting from Vehicle. Initializing these properties and calling the constructor of the base class to initialize the common vehicle properties are done by the Car constructor. To show the details unique to each automobile as well as the common vehicle details, the DisplayDetails() method in the automobile class is overridden.