define a class for rental agency and
The RentalAgency class, which is introduced by the provided C# code, improves the car rental management system by managing a fleet of vehicles using a list and keeping track of the total amount of money made from rentals. Methods for adding cars to the fleet and renting cars for a given number of days depending on their model name are included in the RentalAgency class. When a car is hired, it is taken out of the fleet and the rental amount is multiplied by the number of days to update the overall revenue. An appropriate message is shown if the requested vehicle cannot be located. The Motorcycle class overrides the DisplayDetails() function to add motorcycle-specific details in addition to generic vehicle details.